fixing additional review comments

+++ packages/contracts/contracts/interfaces/v0.8.x/ISharedMinterDAExpV0.sol
@@ -7,10 +7,10 @@ interface ISharedMinterDAExpV0 {
     event SetAuctionDetailsExp(
         uint256 indexed _projectId,
         address indexed _coreContract,
-        uint256 _auctionTimestampStart,
-        uint256 _priceDecayHalfLifeSeconds,
-        uint256 _startPrice,
-        uint256 _basePrice
+        uint64 _auctionTimestampStart,
+        uint64 _priceDecayHalfLifeSeconds,
+        uint128 _startPrice,
+        uint128 _basePrice
     );
     /// Maximum and minimum allowed price decay half lifes updated.
     event AuctionMinHalfLifeSecondsUpdated(
